requirements.txt
chromadb>=0.4.22
sentence-transformers>=2.2.2
PyYAML>=6.0
tiktoken>=0.5.0
